Debug

+++ dataset/vevo_dataset.py
@@ -306,7 +306,7 @@ class VevoDataset(Dataset):
         elif "minor" in key_v2:
             feature_key_v2 = torch.tensor([key_dic["A minor"]])        
         
-        feature_key_v2 = feature_key_v2.to(torch.long)
+        feature_key_v2 = feature_key_v2.to(torch.float)
 
         feature_chord = torch.from_numpy(feature_chord)
         feature_chord = feature_chord.to(torch.long)
